requirements.txt
Flask~=3.1.0
openai~=1.77.0
beautifulsoup4~=4.13.4
requests~=2.32.3
python-dotenv~=1.1.0